Add header toolbar
To make B3 a bit more app like, a toolbar is introduced. It has the same
color as the web app manifest, to blend in nicely on phones.

By doing this, all save buttons should be placed here instead.
Therefore, some CSS is updated to change button presentation.

The grid is also updated to use non-fixed positioning, and
auto-scrolling for the main area, now expected to be contained in a
<main> element.